// src/app/models/calendar.ts
export interface Calendar {
    id?: string;
    uid?: string;
    title?: string;
    emisorName?: string;
    emisorUid?: string;
    desc?: string;
    cupos?: string;
    catId?: string;
    startTime?: Date;
    endTime?: Date;
    allDay?: boolean;
    price?: string;
    product?: string;
    Date?: string;
}